1) Removed some calls to fixup_scopes() since these should no longer be    necessary (because of the hook which calls fixup_scopes() automatically    at tree-insertion time). 2) Added extensive debugging printouts to flatten.k and simple.k 3) Added checks to make sure flattening and simplification actually worked. 4) Fixed some dpl types in header.pro that were overly restrictive. 5) Delayed removal of primed operators until type checking is no longer    done. 6) Totally re-wrote let simplification to deal correctly with changing    scopes.  This was needed to get qsort.pro to transform.
qsort.pro transforms now, but the output won't compile.  I need to get type
checking working for Insert/Extract before I can look for other bugs in the
transformations.  Although the type inference engine should be able to deal
with these functions (it was designed to do so), it has never been hooked
up to the high-level (e.g., Proteus language) code and tested.  Instead of
using these type inference features to typecheck these functions, I'm going
to try to use the backdoor I installed for the typechecking of the
higher-order primed functions.  This should be easier to get working, and
still makes use of the type inference engine, it just bypasses the
arithmetic type clauses.


SKIPPED:
	libmaa/version.c